[REM] l10n_ro_cpv_code:  Merge into l10n_ro_edi
In this commit:
- we merge the l10n_ro_cpv_code module into l10n_ro_edi

task-5416833
