+++ .weblate.json
@@ -2625,12 +2625,6 @@
                 "language_regex": "^(ro)$"
             },
             {
-                "name": "l10n_ro_cpv_code",
-                "filemask": "addons/l10n_ro_cpv_code/i18n/*.po",
-                "new_base": "addons/l10n_ro_cpv_code/i18n/l10n_ro_cpv_code.pot",
-                "language_regex": "^(ro)$"
-            },
-            {
                 "name": "l10n_ro_edi",
                 "filemask": "addons/l10n_ro_edi/i18n/*.po",
                 "new_base": "addons/l10n_ro_edi/i18n/l10n_ro_edi.pot",
