--- addons/l10n_ro_cpv_code/__init__.py
@@ -1,1 +1,0 @@
-from . import models
