--- addons/l10n_ro_cpv_code/models/__init__.py
@@ -1,3 +1,0 @@
-from . import account_edi_xml_ubl_ciusro
-from . import cpv_code
-from . import product
