--- addons/l10n_ro_cpv_code/security/ir.model.access.csv
@@ -1,2 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_cpv_code_user,cpv.code.user,model_l10n_ro_cpv_code,base.group_user,1,0,0,0
