+++ addons/l10n_ro_edi/models/__init__.py
@@ -2,6 +2,8 @@ from . import account_edi_xml_ubl_ciusro
 from . import account_move
 from . import account_move_send
 from . import ciusro_document
+from . import cpv_code
+from . import product
 from . import res_company
 from . import res_config_settings
 from . import res_partner
