+++ addons/l10n_ro_edi/models/account_edi_xml_ubl_ciusro.py
@@ -75,6 +75,17 @@ class AccountEdiXmlUbl_Ro(models.AbstractModel):
 
         return party_node
 
+    def _add_invoice_line_item_nodes(self, line_node, vals):
+        super()._add_invoice_line_item_nodes(line_node, vals)
+
+        product = vals['base_line']['product_id']
+        line_node['cac:Item']['cac:CommodityClassification'] = {
+            'cbc:ItemClassificationCode': {
+                '_text': product.cpv_code_id.code,
+                'listID': 'CPV',
+            }
+        }
+
     # -------------------------------------------------------------------------
     # EXPORT: Constraints
     # -------------------------------------------------------------------------
