+++ addons/l10n_ro_edi/security/ir.model.access.csv
@@ -1,3 +1,4 @@
 id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
 l10n_ro_edi.access_l10n_ro_edi_document_readonly,access_l10n_ro_edi_document_readonly,l10n_ro_edi.model_l10n_ro_edi_document,base.group_user,1,0,0,0
 l10n_ro_edi.access_l10n_ro_edi_document_group_invoice,access_l10n_ro_edi_document_group_invoice,l10n_ro_edi.model_l10n_ro_edi_document,account.group_account_invoice,1,1,1,1
+access_cpv_code_user,cpv.code.user,model_l10n_ro_cpv_code,base.group_user,1,0,0,0
